requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
google-genai>=1.39.0
python-dotenv>=1.0.0

